Add v1.5.19 level engine support, external camera, debug graphs, and multiple fixes
Engine support for v1.5.19-format levels: CLevel:init() compatibility shim,
CLuaVector method accessors, CRGBA alpha default, module texture support,
gate system (AABB trigger volumes), and rewritten server helpers.lua bridging
collision callbacks to v1.5.19-style entity methods.

Refactored external camera into standalone CExternalCameraTask with entity-following
mode, PIP sky rendering, and configurable viewport. Added real-time debug graphs
(FPS, ping, interpolation keys, LCT, delta time) with F4 cycling through modes.

Replay system: keyboard speed controls (pause, slow-mo, fast-forward, reverse).
Music controls on F5/F6/Shift+F7 replacing crash-causing bot keys.

Bug fixes: intermittent scoring failure from lost Lua metatables (lunar.h),
team assignment caching for consistent results, team score flickering from
per-frame reset, server update order for paint-level scoring, negative score
support, bot replay validation for team start positions.

Other: login Enter key support, HUD author line fix, 27 chat emojis,
64-bit compiler warning fixes, post-build script copies level/lua files.

+++ client/src/gui_custom.cpp
@@ -117,8 +117,8 @@ bool CGuiCustom::load(const string &path)
 	for(i=0; i<files.size(); i++)
 	{
 		string fn = files[i];
-		uint32 p = files[i].find(".toc");
-		uint32 pe = files[i].size()-4;
+		uint32 p = (uint32)files[i].find(".toc");
+		uint32 pe = (uint32)files[i].size()-4;
 		if(p==pe)
 		{
 			tocFiles.push_back(files[i]);
